Update trusted-repositories.csv
Proposed adding the World Bank's Reproducible Research Repository. Updated the link to the Microdata Library preservation policy (which was broken). The Microdata Library and the Reproducible Research Repository follow the Records Retention and Disposition Policy (RRDP) which is set at the level of the World Bank's Development Economics Vice Presidency. The only publicly-available documentation of the RRDP is the admin manual, which I have linked. The policy mandates we permanently retain anonymized data and  research products.